backend/requirements.txt
fastapi
uvicorn[standard]
motor
httpx
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
python-multipart
